quant_backtesting_framework/risk_management.py
import numpy as np


class RiskManagement:
    """
    The RiskManagement class provides methods for calculating various risk measures.

    Attributes:
    portfolio_returns (numpy.array): The portfolio returns to calculate risk for.
    """
    def __init__(self, ewma_alpha=0.94):
        self.ewma_alpha = ewma_alpha
        self.volatility = None

    def assess_trade_risk(self, portfolio, proposed_trade):
        """
        Assess if a proposed trade complies with the risk parameters.
        """
        # Implement logic to evaluate trade risk here
        # For example, check if the trade exceeds the maximum position size
        # return is_within_risk_limits
        raise NotImplementedError

    def update_volatility(self, returns):
        """
        Update the volatility estimate using EWMA.
        """
        if self.volatility is None:
            self.volatility = returns.std()
        else:
            self.volatility = self.ewma_alpha * returns.std() + (1 - self.ewma_alpha) * self.volatility

    def evaluate_portfolio_risk(self, portfolio):
        """
        Evaluates the overall risk of the portfolio.
        """
        # Implement logic to evaluate portfolio risk
        # For example, calculate portfolio drawdown
        #return is_within_risk_limits
        raise NotImplementedError
    
    def calculate_max_drawdown(self, portfolio_value):
        """
        Calculate the maximum drawdown of the portfolio.

        Args:
        portfolio_value (numpy.array): The portfolio value.

        Returns:
        float: The maximum drawdown.
        """
        # Calculate the running maximum
        running_max = np.maximum.accumulate(portfolio_value)
        # Calculate the drawdown
        drawdown = (running_max - portfolio_value) / running_max
        # Return the maximum drawdown
        return drawdown.max()
    
    def calculate_var(self, returns, confidence_level):
        """
        Calculate the Value at Risk (VaR) at the given confidence level.

        Args:
        confidence_level (float): The confidence level.

        Returns:
        float: The VaR at the given confidence level.
        """
        # Calculate the VaR
        var = np.percentile(returns, 100 * (1 - confidence_level))
        return var
    
    def calculate_cvar(self, returns, confidence_level):
        """
        Calculate the Conditional Value at Risk (CVaR) at the given confidence level.

        Args:
        confidence_level (float): The confidence level.

        Returns:
        float: The CVaR at the given confidence level.
        """
        # Calculate the VaR
        var = self.calculate_var(returns, confidence_level)

        # Calculate the CVaR
        cvar = returns[returns <= var].mean()
        return cvar

if __name__ == '__main__':
    # Assume we have the following portfolio returns
    portfolio_returns = np.array([0.01, -0.02, 0.03, -0.04, 0.05, -0.06, 0.07, -0.08, 0.09, -0.10])

    # Create a RiskManagement instance
    risk_manager = RiskManagement()

    # Calculate the Value at Risk (VaR) at a 95% confidence level
    var = risk_manager.calculate_var(portfolio_returns, 0.95)
    print(f"Value at Risk (95% confidence level): {var}")

    # Calculate the Conditional Value at Risk (CVaR) at a 95% confidence level
    cvar = risk_manager.calculate_cvar(portfolio_returns, 0.95)
    print(f"Conditional Value at Risk (95% confidence level): {cvar}")

    # Assume we have the following portfolio values
    portfolio_values = np.array([100, 110, 105, 120, 115, 130, 125, 140, 135, 150])

    # Calculate the maximum drawdown
    max_drawdown = risk_manager.calculate_max_drawdown(portfolio_values)
    print(f"Maximum Drawdown: {max_drawdown}")
